# backend/openapi.json
{"openapi": "3.1.0", "info": {"title": "BillLens API", "version": "0.1.0"}, "paths": {"/health": {"get": {"summary": "Health Check", "operationId": "health_check_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/samples": {"get": {"summary": "Get Samples", "operationId": "get_samples_samples_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/Sample"}, "type": "array", "title": "Response Get Samples Samples Get"}}}}}}}, "/policies": {"post": {"summary": "Ingest Policy", "operationId": "ingest_policy_policies_post", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/policies/{id}": {"get": {"summary": "Get Policy", "operationId": "get_policy_policies__id__get", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "title": "Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/policies/{id}/terms": {"put": {"summary": "Update Policy Terms", "operationId": "update_policy_terms_policies__id__terms_put", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "title": "Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PolicyTermsUpdate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ResolvedPolicy"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/bills": {"post": {"summary": "Upload Bill", "operationId": "upload_bill_bills_post", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/runs": {"post": {"summary": "Create Run", "operationId": "create_run_runs_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/RunRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/runs/{id}": {"get": {"summary": "Get Run", "operationId": "get_run_runs__id__get", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "title": "Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/runs/{id}/confirm": {"put": {"summary": "Confirm Run", "operationId": "confirm_run_runs__id__confirm_put", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "title": "Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConfirmRunRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/runs/{id}/stream": {"get": {"summary": "Run Stream", "operationId": "run_stream_runs__id__stream_get", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "title": "Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/runs/{id}/events": {"get": {"summary": "Get Run Events", "operationId": "get_run_events_runs__id__events_get", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "title": "Id"}}, {"name": "since", "in": "query", "required": false, "schema": {"type": "integer", "default": 0, "title": "Since"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/runs/{id}/result": {"get": {"summary": "Get Run Result", "operationId": "get_run_result_runs__id__result_get", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "title": "Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RunResult"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/demo/{sample_id}": {"post": {"summary": "Run Demo", "operationId": "run_demo_demo__sample_id__post", "parameters": [{"name": "sample_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Sample Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"ActionPack": {"properties": {"desk_questions": {"items": {"$ref": "#/components/schemas/DeskQuestion"}, "type": "array", "title": "Desk Questions"}, "letter": {"$ref": "#/components/schemas/Letter"}, "checklist": {"items": {"type": "string"}, "type": "array", "title": "Checklist"}}, "type": "object", "required": ["desk_questions", "letter", "checklist"], "title": "ActionPack"}, "BillLine": {"properties": {"line_no": {"type": "integer", "title": "Line No"}, "raw_text": {"type": "string", "title": "Raw Text"}, "qty": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Qty"}, "unit_price": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Unit Price"}, "amount": {"type": "number", "title": "Amount"}, "service_date": {"anyOf": [{"type": "string", "format": "date"}, {"type": "null"}], "title": "Service Date"}, "category_guess": {"anyOf": [{"type": "string", "enum": ["room", "nursing", "professional_fee", "procedure", "pharmacy", "consumable", "investigation", "implant", "other"]}, {"type": "null"}], "title": "Category Guess"}}, "type": "object", "required": ["line_no", "raw_text", "amount"], "title": "BillLine"}, "ConfirmRunRequest": {"properties": {"lines": {"items": {"$ref": "#/components/schemas/MatchedLine"}, "type": "array", "title": "Lines"}, "is_insured": {"type": "boolean", "title": "Is Insured"}}, "type": "object", "required": ["lines", "is_insured"], "title": "ConfirmRunRequest"}, "DeskQuestion": {"properties": {"priority": {"type": "integer", "title": "Priority"}, "question": {"type": "string", "title": "Question"}, "why": {"type": "string", "title": "Why"}, "line_ids": {"items": {"type": "integer"}, "type": "array", "title": "Line Ids", "default": []}, "amount": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Amount"}}, "type": "object", "required": ["priority", "question", "why"], "title": "DeskQuestion"}, "Evidence": {"properties": {"id": {"type": "string", "title": "Id"}, "type": {"type": "string", "enum": ["line", "clause", "catalog", "calc", "kb"], "title": "Type"}, "text": {"type": "string", "title": "Text"}, "meta": {"additionalProperties": true, "type": "object", "title": "Meta", "default": {}}}, "type": "object", "required": ["id", "type", "text"], "title": "Evidence"}, "ExtractedBill": {"properties": {"hospital_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Hospital Name"}, "bill_date": {"anyOf": [{"type": "string", "format": "date"}, {"type": "null"}], "title": "Bill Date"}, "stated_total": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Stated Total"}, "estimate_amount": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Estimate Amount"}, "lines": {"items": {"$ref": "#/components/schemas/BillLine"}, "type": "array", "title": "Lines"}, "notes": {"items": {"type": "string"}, "type": "array", "title": "Notes", "default": []}}, "type": "object", "required": ["lines"], "title": "ExtractedBill"}, "Finding": {"properties": {"id": {"type": "string", "title": "Id"}, "rule_code": {"type": "string", "title": "Rule Code"}, "kind": {"type": "string", "enum": ["billing", "coverage"], "title": "Kind"}, "severity": {"type": "string", "enum": ["red", "amber", "info"], "title": "Severity"}, "origin": {"type": "string", "enum": ["rule", "agent"], "title": "Origin"}, "line_ids": {"items": {"type": "integer"}, "type": "array", "title": "Line Ids", "default": []}, "title": {"type": "string", "title": "Title"}, "explanation": {"type": "string", "title": "Explanation"}, "amount_at_stake": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Amount At Stake"}, "confidence": {"type": "string", "enum": ["high", "medium", "low"], "title": "Confidence"}, "evidence_ids": {"items": {"type": "string"}, "type": "array", "title": "Evidence Ids", "default": []}, "suggested_question": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Suggested Question"}, "uncertain": {"type": "boolean", "title": "Uncertain", "default": false}, "agent_note": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Agent Note"}, "dismissed": {"type": "boolean", "title": "Dismissed", "default": false}}, "type": "object", "required": ["id", "rule_code", "kind", "severity", "origin", "title", "explanation", "confidence"], "title": "Finding"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "Letter": {"properties": {"subject": {"type": "string", "title": "Subject"}, "body": {"type": "string", "title": "Body"}, "generated_by": {"type": "string", "enum": ["template", "llm"], "title": "Generated By"}}, "type": "object", "required": ["subject", "body", "generated_by"], "title": "Letter"}, "MatchedLine": {"properties": {"line_no": {"type": "integer", "title": "Line No"}, "raw_text": {"type": "string", "title": "Raw Text"}, "qty": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Qty"}, "unit_price": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Unit Price"}, "amount": {"type": "number", "title": "Amount"}, "service_date": {"anyOf": [{"type": "string", "format": "date"}, {"type": "null"}], "title": "Service Date"}, "category_guess": {"anyOf": [{"type": "string", "enum": ["room", "nursing", "professional_fee", "procedure", "pharmacy", "consumable", "investigation", "implant", "other"]}, {"type": "null"}], "title": "Category Guess"}, "catalog_id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Catalog Id"}, "canonical_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Canonical Name"}, "category": {"anyOf": [{"type": "string", "enum": ["room", "nursing", "professional_fee", "procedure", "pharmacy", "consumable", "investigation", "implant", "other"]}, {"type": "null"}], "title": "Category"}, "match_method": {"type": "string", "enum": ["exact", "embedding", "llm", "unmatched"], "title": "Match Method", "default": "unmatched"}, "match_score": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Match Score"}, "extraction_flag": {"type": "boolean", "title": "Extraction Flag", "default": false}, "user_edited": {"type": "boolean", "title": "User Edited", "default": false}}, "type": "object", "required": ["line_no", "raw_text", "amount"], "title": "MatchedLine"}, "PolicyTerm": {"properties": {"key": {"type": "string", "enum": ["sum_insured", "room_rent_cap", "copay_pct", "proportionate_deduction"], "title": "Key"}, "status": {"type": "string", "enum": ["extracted", "manual", "not_found"], "title": "Status"}, "value": {"anyOf": [{"type": "number"}, {"type": "boolean"}, {"type": "null"}], "title": "Value"}, "room_cap_kind": {"anyOf": [{"type": "string", "enum": ["absolute", "pct_si"]}, {"type": "null"}], "title": "Room Cap Kind"}, "quote": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Quote"}, "page": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Page"}, "chunk_id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Chunk Id"}, "similarity": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Similarity"}}, "type": "object", "required": ["key", "status"], "title": "PolicyTerm"}, "PolicyTermsUpdate": {"properties": {"terms": {"additionalProperties": true, "type": "object", "title": "Terms"}, "confirmed": {"type": "boolean", "title": "Confirmed"}}, "type": "object", "required": ["terms", "confirmed"], "title": "PolicyTermsUpdate"}, "ResolvedPolicy": {"properties": {"is_insured": {"type": "boolean", "title": "Is Insured"}, "sum_insured": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Sum Insured"}, "room_rent_cap_per_day": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Room Rent Cap Per Day"}, "copay_pct": {"type": "number", "title": "Copay Pct", "default": 0.0}, "proportionate_deduction": {"type": "boolean", "title": "Proportionate Deduction", "default": false}, "term_evidence": {"additionalProperties": {"type": "string"}, "type": "object", "title": "Term Evidence", "default": {}}}, "type": "object", "required": ["is_insured"], "title": "ResolvedPolicy"}, "RunRequest": {"properties": {"bill_id": {"type": "string", "title": "Bill Id"}, "policy_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Policy Id"}, "auto_confirm": {"anyOf": [{"type": "boolean"}, {"type": "null"}], "title": "Auto Confirm"}}, "type": "object", "required": ["bill_id"], "title": "RunRequest"}, "RunResult": {"properties": {"run_id": {"type": "string", "title": "Run Id"}, "bill": {"$ref": "#/components/schemas/ExtractedBill"}, "lines": {"items": {"$ref": "#/components/schemas/MatchedLine"}, "type": "array", "title": "Lines"}, "policy_terms": {"additionalProperties": {"$ref": "#/components/schemas/PolicyTerm"}, "type": "object", "title": "Policy Terms", "default": {}}, "policy": {"$ref": "#/components/schemas/ResolvedPolicy"}, "findings": {"items": {"$ref": "#/components/schemas/Finding"}, "type": "array", "title": "Findings"}, "evidence": {"items": {"$ref": "#/components/schemas/Evidence"}, "type": "array", "title": "Evidence"}, "simulator": {"$ref": "#/components/schemas/SimulatorResult"}, "summary": {"$ref": "#/components/schemas/Summary"}, "action_pack": {"anyOf": [{"$ref": "#/components/schemas/ActionPack"}, {"type": "null"}]}, "mode": {"type": "string", "enum": ["live", "replay"], "title": "Mode", "default": "live"}}, "type": "object", "required": ["run_id", "bill", "lines", "policy", "findings", "evidence", "simulator", "summary"], "title": "RunResult"}, "Sample": {"properties": {"id": {"type": "string", "title": "Id"}, "kind": {"type": "string", "title": "Kind"}, "title": {"type": "string", "title": "Title"}, "description": {"type": "string", "title": "Description"}, "thumb_url": {"type": "string", "title": "Thumb Url"}, "is_insured": {"type": "boolean", "title": "Is Insured"}}, "type": "object", "required": ["id", "kind", "title", "description", "thumb_url", "is_insured"], "title": "Sample"}, "SimulatorResult": {"properties": {"applied": {"type": "boolean", "title": "Applied"}, "billed_total": {"type": "number", "title": "Billed Total"}, "non_payable_total": {"type": "number", "title": "Non Payable Total", "default": 0}, "room_rent_deduction": {"type": "number", "title": "Room Rent Deduction", "default": 0}, "copay_amount": {"type": "number", "title": "Copay Amount", "default": 0}, "insurer_pays": {"type": "number", "title": "Insurer Pays"}, "patient_pays": {"type": "number", "title": "Patient Pays"}, "steps": {"items": {"$ref": "#/components/schemas/WaterfallStep"}, "type": "array", "title": "Steps", "default": []}, "assumptions": {"items": {"type": "string"}, "type": "array", "title": "Assumptions", "default": []}}, "type": "object", "required": ["applied", "billed_total", "insurer_pays", "patient_pays"], "title": "SimulatorResult"}, "Summary": {"properties": {"billed_total": {"type": "number", "title": "Billed Total"}, "questionable_total": {"type": "number", "title": "Questionable Total"}, "insurer_deductions_total": {"type": "number", "title": "Insurer Deductions Total"}, "insurer_pays": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Insurer Pays"}, "patient_pays": {"type": "number", "title": "Patient Pays"}, "counts": {"additionalProperties": {"type": "integer"}, "type": "object", "title": "Counts"}, "agent_stats": {"additionalProperties": {"type": "integer"}, "type": "object", "title": "Agent Stats", "default": {}}}, "type": "object", "required": ["billed_total", "questionable_total", "insurer_deductions_total", "insurer_pays", "patient_pays", "counts"], "title": "Summary"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}, "WaterfallStep": {"properties": {"key": {"type": "string", "enum": ["billed", "non_payable", "room_rent", "copay", "sum_insured_cap", "insurer_pays", "patient_pays"], "title": "Key"}, "label": {"type": "string", "title": "Label"}, "amount": {"type": "number", "title": "Amount"}, "running_total": {"type": "number", "title": "Running Total"}, "explanation": {"type": "string", "title": "Explanation"}, "evidence_ids": {"items": {"type": "string"}, "type": "array", "title": "Evidence Ids", "default": []}}, "type": "object", "required": ["key", "label", "amount", "running_total", "explanation"], "title": "WaterfallStep"}}}}